k8s/snapshot/es-snapshot/requirements.txt
elasticsearch
boto3
python_dateutil
